+++ plugins/woocommerce-admin/client/analytics/components/report-chart/index.js
@@ -42,7 +42,8 @@ export class ReportChart extends Component {
 					const label = intervalData[ segment.segment_label ]
 						? segment.segment_label + ' (#' + segment.segment_id + ')'
 						: segment.segment_label;
-					intervalData[ label ] = {
+					intervalData[ segment.segment_id ] = {
+						label,
 						value: segment.subtotals[ selectedChart.key ] || 0,
 					};
 				}
@@ -59,8 +60,6 @@ export class ReportChart extends Component {
 		const { query, primaryData, secondaryData, selectedChart } = this.props;
 		const currentInterval = getIntervalForQuery( query );
 		const { primary, secondary } = getCurrentDates( query );
-		const primaryKey = `${ primary.label } (${ primary.range })`;
-		const secondaryKey = `${ secondary.label } (${ secondary.range })`;
 
 		const chartData = primaryData.data.intervals.map( function( interval, index ) {
 			const secondaryDate = getPreviousDate(
@@ -74,11 +73,13 @@ export class ReportChart extends Component {
 			const secondaryInterval = secondaryData.data.intervals[ index ];
 			return {
 				date: formatDate( 'Y-m-d\\TH:i:s', interval.date_start ),
-				[ primaryKey ]: {
+				primary: {
+					label: `${ primary.label } (${ primary.range })`,
 					labelDate: interval.date_start,
 					value: interval.subtotals[ selectedChart.key ] || 0,
 				},
-				[ secondaryKey ]: {
+				secondary: {
+					label: `${ secondary.label } (${ secondary.range })`,
 					labelDate: secondaryDate.format( 'YYYY-MM-DD HH:mm:ss' ),
 					value: ( secondaryInterval && secondaryInterval.subtotals[ selectedChart.key ] ) || 0,
 				},